// src/ui-kit/Basics/types.ts
type WithBreakpointArgs<Props> = {
  [Key in keyof Props]?: Props[Key];
};

interface BoxPropsBase {
  /* See props documentation at:
   * https://www.smooth-code.com/open-source/smooth-ui/docs/box/#box-2
   */
  background: number | string;
  backgroundColor: string;
  backgroundImage: number | string;
  backgroundSize: number | string;
  backgroundPosition: number | string;
  backgroundRepeat: number | string;
  opacity: number | string;
  overflow: number | string;
  transition: number | string;
  border: number | string;
  borderTop: number | string;
  borderTopColor: number | string;
  borderRight: number | string;
  borderRightColor: number | string;
  borderBottom: number | string;
  borderBottomColor: number | string;
  borderLeft: number | string;
  borderLeftColor: number | string;
  borderColor: number | string;
  borderWidth: number | string;
  borderStyle: number | string;
  borderRadius: string;
  display: number | string;
  alignItems: number | string;
  alignContent: number | string;
  justifyContent: number | string;
  justifyItems: number | string;
  flexWrap: number | string;
  flexBasis: number | string;
  flexDirection: number | string;
  flex: number | string;
  justifySelf: number | string;
  alignSelf: number | string;
  order: number | string;
  gridGap: number | string;
  gridColumnGap: number | string;
  gridRowGap: number | string;
  gridColumn: number | string;
  gridRow: number | string;
  gridAutoFlow: number | string;
  gridAutoColumns: number | string;
  gridAutoRows: number | string;
  gridTemplateColumns: number | string;
  gridTemplateRows: number | string;
  gridTemplateAreas: number | string;
  gridArea: number | string;
  width: number | string;
  height: number | string;
  maxWidth: number | string;
  maxHeight: number | string;
  minWidth: number | string;
  minHeight: number | string;
  size: number | string;
  verticalAlign: number | string;
  position: number | string;
  zIndex: number | string;
  top: number | string;
  right: number | string;
  bottom: number | string;
  left: number | string;
  boxShadow: string;
  textShadow: string;
  margin: number | string;
  m: number | string;
  marginTop: number | string;
  mt: number | string;
  marginRight: number | string;
  mr: number | string;
  marginBottom: number | string;
  mb: number | string;
  marginLeft: number | string;
  ml: number | string;
  mx: number | string;
  my: number | string;
  padding: number | string;
  p: number | string;
  paddingTop: number | string;
  pt: number | string;
  paddingRight: number | string;
  pr: number | string;
  paddingBottom: number | string;
  pb: number | string;
  paddingLeft: number | string;
  pl: number | string;
  px: number | string;
  py: number | string;
  fontFamily: string;
  fontSize: number | string;
  lineHeight: number | string;
  fontWeight: number | string;
  textAlign: number | string;
  letterSpacing: number | string;
  color: string;
  textTransform: number | string;
  row: number | string | boolean;
  col: number | string | boolean;
  disabled: boolean;
  cursor: string;
  bg: string;
  transform: string | number;
}

export type BoxProps = WithBreakpointArgs<BoxPropsBase>;